# legacy/brain/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for Arthemis Brain.",
        "title": "Arthemis Brain API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/proponent/": {
            "get": {
                "description": "Retrieves every proponent",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proponent"
                ],
                "summary": "Gets all proponents",
                "responses": {
                    "200": {
                        "description": "Proponent deleted successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "404": {
                        "description": "Proponent not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proponent/create": {
            "post": {
                "description": "Creates a new proponent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proponent"
                ],
                "summary": "Create Proponent",
                "parameters": [
                    {
                        "description": "Proponent details",
                        "name": "proponent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Proponent"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Proponent created successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or bad request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proponent/delete/{id}": {
            "delete": {
                "description": "Deletes a proponent by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proponent"
                ],
                "summary": "Delete Proponent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proponent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proponent deleted successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "ProponentID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Proponent not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proponent/update/{id}": {
            "patch": {
                "description": "Updates an existing proponent by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proponent"
                ],
                "summary": "Update Proponent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proponent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Proponent details to update",
                        "name": "proponent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Proponent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Proponent"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or ProponentID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Proponent not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/proponent/{id}": {
            "get": {
                "description": "Retrieves a proponent by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proponent"
                ],
                "summary": "Get Proponent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proponent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Proponent"
                        }
                    },
                    "400": {
                        "description": "ProponentID not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Proponent not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Activity": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Indicator"
                    }
                },
                "justification": {
                    "type": "string"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Location"
                    }
                },
                "name": {
                    "type": "string"
                },
                "projectID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Indicator": {
            "type": "object",
            "properties": {
                "activityID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "justification": {
                    "type": "string"
                },
                "locationID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "observationMethod": {
                    "type": "string"
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Observation"
                    }
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "valueBaseline": {
                    "type": "number"
                },
                "valueReference": {
                    "type": "number"
                }
            }
        },
        "models.JSONB": {
            "type": "object",
            "additionalProperties": {}
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "ecosystem": {
                    "type": "string"
                },
                "extent": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Indicator"
                    }
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "projectID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Observation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "indicatorID": {
                    "type": "integer"
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "models.Project": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Activity"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "justification": {
                    "type": "string"
                },
                "lifetimeEnd": {
                    "type": "string"
                },
                "lifetimeStart": {
                    "type": "string"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Location"
                    }
                },
                "name": {
                    "type": "string"
                },
                "projectProponents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProjectProponent"
                    }
                },
                "proponentID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ProjectProponent": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "projectID": {
                    "type": "integer"
                },
                "proponentID": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Proponent": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Project"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Tarefa não encontrada"
                },
                "message": {
                    "type": "string",
                    "example": "ID inválido fornecido"
                }
            }
        }
    }
}